ai/requirements.txt
google-generativeai
ipython
Pillow